==> src/notebooks/Owen Corrigan/02-oc-extract_keyframes.ipynb <==
from pathlib import Path

from tqdm import tqdm_notebook as tqdm

%%time

images_base = Path('../data/interim/images/dev-set')
images_base.mkdir(parents=True, exist_ok=True)
videos = sorted(Path('../data/raw/Memorability 2018/dev-set/sources').iterdir())
times = ['00:00:00', '00:00:03.5', '00:00:06.958']
frames = [0, 84, 168]
output_file = 'out.png'

for video in videos:
    for time, frame in zip(times, frames):
        output_file = '{}-{}.png'.format((images_base / video.stem), frame)
        if not Path(output_file).exists():
            ffmpeg_args = ['ffmpeg', '-ss', time, '-i', str(video), '-f', 'image2', '-vframes', '1', output_file, '-y']
            result = subprocess.Popen(args=ffmpeg_args, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
            out, err = result.communicate()

   input_file = './data/raw/Memorability 2018/dev-set/sources/video10.webm'         if result.returncode:
                raise Exception("Non zero exit code")

%%time

images_base = Path('./data/interim/test-set/images')
images_base.mkdir(parents=True, exist_ok=True)
videos = sorted(Path('./data/raw/Memorability 2018/test-set/sources').iterdir())
times = ['00:00:00', '00:00:03.5', '00:00:06.958']
frames = [0, 84, 168]
output_file = 'out.png'

for video in tqdm(videos):
    for time, frame in zip(times, frames):
        output_file = '{}-{}.png'.format((images_base / video.stem), frame)
        if not Path(output_file).exists():
            ffmpeg_args = ['ffmpeg', '-ss', time, '-i', str(video), '-f', 'image2', '-vframes', '1', output_file, '-y']
            result = subprocess.Popen(args=ffmpeg_args, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
            out, err = result.communicate()

            if result.returncode:
                raise Exception("Non zero exit code")
