==> src/notebooks/Owen Corrigan/04-oc-experiments.ipynb <==
import json
import socket
import time

import click

import pandas as pd
import numpy as np
import tensorflow as tf

from tqdm import tqdm_notebook as tqdm

from src.models import cli
from src.models.inception_model import InceptionModel

# Train Model From Notebook

experiment_number = 2
experiment_name = "blah"
batches = 1000
report_step = 200
disable_probress_bar = False
max_train_time_seconds = None

meta_folder = 'reports/experiments/{:03}-{}/'
meta_folder = meta_folder.format(experiment_number,
                                 experiment_name)

model = InceptionModel(batch_size=32, meta_folder=meta_folder, term='long-term')
model.train(batches, report_step, disable_progress_bar=disable_probress_bar,
                     max_train_time_seconds=max_train_time_seconds)
model.eval()

# Find Standard Deviation of Dummy Pearson R

predictions, ground_truth = model.predict(include_ground_truth=True)

pearson_values = []
for i in tqdm(range(1000)):
    rand = np.random.random(size=len(ground_truth))
    pearson_values += [pd.DataFrame([rand, ground_truth]).T.corr().iloc[1, 0]]
pearson_values = np.array(pearson_values)
print(pearson_values.mean())
print(pearson_values.std())
