==> src/notebooks/Owen Corrigan/05-oc-submissions.ipynb <==
import pandas as pd

# load ground truth
ground_truth_file = './data/raw/Memorability 2018/dev-set/ground-truth/ground-truth_dev-set.csv'
ground_truth = pd.read_csv(ground_truth_file).set_index('video')
short_term = ground_truth['short-term_memorability']
long_term = ground_truth['long-term_memorability']

# create videos
captions = open('./data/raw/Memorability 2018/test-set/test-set_videos-captions.txt', 'r').read()
videos = [i.split('\t')[0] for i in captions.split('\n')]
results = pd.DataFrame(videos, columns=['video'])

# 
cols = ['video', 'score', 'confidence']
long_term_results = results.assign(score=long_term.mean(), confidence=0).sort_values('video')[cols]
long_term_results.to_csv('./reports/me18n_insight@dcu_longterm_run1-mean.csv', header=False, index=False)

short_term_results = results.assign(score=short_term.mean(), confidence=0).sort_values('video')[cols]
short_term_results.head()
short_term_results.to_csv('./reports/me18n_insight@dcu_shorterm_run1-mean.csv', header=False, index=False)
