==> src/notebooks/Owen Corrigan/06-oc-aesthetic-analysis.ipynb <==
import json
import pickle

import pandas as pd

# load aesthetic features
aesthethic = pickle.load(open('./data/external/aesthetic/aesthetic_score.pkl', 'rb'))
aesthethic = pd.DataFrame([(i, j[0]) for i, j in aesthethic])

video = aesthethic[0].str.split('-').apply(lambda x: x[0]+'.webm')
frame = aesthethic[0].str.split('-').apply(lambda x: int(x[1].split('.png')[0]))
aesthethic['video'] = video
aesthethic['frame'] = frame
aesthethic = aesthethic.drop(0, axis=1).rename(columns={1: 'score'})
aesthethic = pd.DataFrame(aesthethic.groupby('video')['score'].mean())

aesthethic.head()

# combine with ground truth
annotations = json.load(open('./data/processed/annotations.json'))
test_annotations = aesthethic.loc[annotations['test']]

ground_truth_file = './data/raw/Memorability 2018/dev-set/ground-truth/ground-truth_dev-set.csv'
ground_truth = pd.read_csv(ground_truth_file).set_index('video').join(aesthethic)
ground_truth.head()

cols = ['short-term_memorability', 'score'] 
spearman = ground_truth.loc[annotations['test']][cols].corr('spearman').iloc[0, 1]
print("Spearman: {:.04}".format(spearman))
