==> src/notebooks/Owen Corrigan/07-oc-saliency.ipynb <==
# Saliency Features

Demonstrate how to use saliency features

import h5py
import numpy as np

path = './data/external/salience/dev-set/features.h5'
hdf5_features = h5py.File(path, "r")
path = './data/external/salience/dev-set/indices.npy'
indices = np.load(path)

features = hdf5_features['features1'][:]
features.shape

Note: test set has 8000 videos, and 3 images per video

# sample of features
features[0]

indices[:10]

Note: features are **not** in incorrect order

# use arg sort to get indices in correct order
np.argsort(indices)

# these are the features we feed to tensorflow
features[np.argsort(indices)[::3]].shape
