==> src/notebooks/Owen Corrigan/08-oc-EEG-features.ipynb <==
import pandas as pd
import json

# EEG Features

Extra layer of complexity, since these features were only extracted for a subset of the whole dataset

eeg = pd.read_csv('./data/external/eeg/dev-set/eeg.csv', header=None)
eeg = (eeg.rename(columns={0: 'video', 1: 'vals'})
        .sort_values('video'))
eeg['video'] = eeg['video'] + '.webm'

eeg.head()

eeg.shape

annotations = json.load(open('data/processed/annotations.json'))

# are all videos in dev-set?
eeg['video'].isin(annotations['dev-set']['all']).all()

# create new annotation mode, which filters dev-set to just those covered by eeg annotations

eeg_train = [i for i in eeg['video'] if i in annotations['dev-set']['train']]
eeg_test = [i for i in eeg['video'] if i in annotations['dev-set']['test']]
eeg_val = [i for i in eeg['video'] if i in annotations['dev-set']['val']]
eeg_all = annotations['dev-set']['all']

annotations['eeg-set'] = {
    'train': eeg_train,
    'test': eeg_test,
    'val': eeg_val,
    'all': eeg_all
}

json.dump(annotations, open('data/processed/annotations.json', 'w'))
