==> src/notebooks/Owen Corrigan/09-oc-ensemble.ipynb <==
import pandas as pd
from pathlib import Path
import json

from sklearn.ensemble import RandomForestRegressor

paths = list(Path('reports/experiments/').glob('**/*.csv'))
paths[:5]

annotations = json.load(open('data/processed/annotations.json'))

# short term predictions

short_paths = [i for i in paths if 'short-term/dev' in str(i)]
short_paths

test_idx = pd.read_csv(short_paths[0], header=None)[0].values

preds = [pd.read_csv(path, header=None)[1] for path in short_paths]
preds = pd.DataFrame(preds).T.assign(video=test_idx).set_index('video')

y = pd.read_csv('data/raw/Memorability 2018/dev-set/ground-truth/ground-truth_dev-set.csv').set_index('video')['short-term_memorability']
y.head()

X_train = preds
y_train = y.loc[annotations['dev-set']['test']]

clf = RandomForestRegressor()
clf.fit(X_test, y_test)

sorted(short_paths)[:6]

short_paths = [i for i in paths if 'short-term/test' in str(i)]
test = [pd.read_csv(path, header=None)[1] for path in sorted(short_paths)[:6]]
test = pd.DataFrame(test).T
test.head()

cols = ['video', 'score', 'confidence']
results = pd.DataFrame(annotations['test-set']['test'], columns=['video'])
results = (results.assign(score=clf.predict(test), confidence=1)
                          .sort_values('video')[cols])

results.to_csv('reports/experiments/009-ensemble-short-term/test-set.csv', header=False, index=False)

# Long term predictions

long_paths = [i for i in paths if 'long-term/dev' in str(i)]
long_paths

test_idx = pd.read_csv(long_paths[0], header=None)[0].values

preds = [pd.read_csv(path, header=None)[1] for path in long_paths]
preds = pd.DataFrame(preds).T.assign(video=test_idx).set_index('video')

y = pd.read_csv('data/raw/Memorability 2018/dev-set/ground-truth/ground-truth_dev-set.csv').set_index('video')['short-term_memorability']
y.head()

X_train = preds
y_train = y.loc[annotations['dev-set']['test']]

clf = RandomForestRegressor()
clf.fit(X_test, y_test)

long_paths = [i for i in paths if 'long-term/test' in str(i)]
test = [pd.read_csv(path, header=None)[1] for path in sorted(long_paths)[:6]]
test = pd.DataFrame(test).T
test.head()

test.shape

cols = ['video', 'score', 'confidence']
results = pd.DataFrame(annotations['test-set']['test'], columns=['video'])
results = (results.assign(score=clf.predict(test), confidence=1)
                          .sort_values('video')[cols])

results.to_csv('reports/experiments/009-ensemble-long-term/test-set.csv', header=False, index=False)
